Website development agency
There is an agency that specialize on creating website/landing on the site. The site includes adaptive versions for different types of devices. For development was used VS code 2019 and html/css/js (JQuery).
This project can be addicted by connection to database and development of backend.
Also, this projet was created by copy-paste way with me. The main source of information was pointed out bellow.
https://www.youtube.com/playlist?list=PLoq3Accf02PVO4GvY4-UtIQkeD6tNmX_f
I thought, that every developer should note the sourse of his knowledge with ethical point of view.